requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# CLI & Display
typer[all]==0.9.0
rich==13.7.0
click==8.1.7
tqdm==4.66.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI & NLP
openai==1.3.7
langchain==0.0.345
langchain-community==0.0.10
faiss-cpu==1.7.4
dashscope==1.14.0
tiktoken==0.5.2

# Document Processing
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Optional: Performance
redis==5.0.1
celery==5.3.4